StateManager owns its states and will ensure they are destroy()ed.
Hopefully, this is yet another resource management improvement: we now
have means to guarantee that resources allocated by a GameState are
freed. (On the other hand, it is no longer safe to keep external
references to GameStates -- the State Manager owns them, and will
destroy them even if someone else has references to it. This shouldn't
be an issue in practice.)