OrgArch: Added LowLevelEventHandler; worked on StateManager. Buggy.
This is a "I did too much so I have to save my work even if it is
buggy" type of commit. Replaced the old 'EventHandler' with a new
'LowLevelEventHandler' (which is basically the same idea, but with a
new name and adapted to the new architecture).

There are bugs (the 'states_simple' example crashes in many
situations) and several debug prints throughout the code.

But there is a lot of work done in what seems to be the right
direction. Ergo, commit!

--HG--
branch : lmb/organize_architecture