OrgArch: Yesterday's bug seem to be fixed. Code still ugly.
Lots of wonderings about what to do when, for instance, a state pops
itself from the stack and still has events for it in the event
queue. Or: When shall a newly pushed state start to get its event
handlers called?

These are delicate things, because any minor problem in the chosen
semantics is very likely to cause bugs that are very hard to
understand and fix. And this includes timing issues ("if this hapens
just before that..."), which are always confusing.

And, finally, there is performance. For FewDee, I am trying to follow
a "convenience over performance" policy, but this doesn't mean that I
shall simply ignore performance issues. In this case (i.e., event
handling), in particular, we are talking about the main game loop... I
should ideally find a reasonably efficient way to implement the
convenient semantics I want.

Anyway... bug seems fixed. The debug prints are still there, lots of
documentation is missing, and a few "deeper" TODOs still need to be
addressed. Yet, another step in what seems to be the right direction,
so, here's a new commit.

And the longest commmit message I write in a long time.

--HG--
branch : lmb/organize_architecture