Smarter encoding of times in events.
Now using just one of the user data fields. In 32-bit architectures,
the precision of the data will be limited to a single-precision
float. This shouldn't be a problem: for deltaTime this is enough
resolution; for totalTime, some precision will be lost in very
long-running programs, but errors will not accumulate, since all the
bookkeeping is still done with doubles (the conversion is made just
when filling the event data).