Config: Added more parser tests (AA, list and CTFE).
Now the parser looks complete and solid. Still need to document a bit
more, but otherwise I am considering it complete.

(Not that the implementation cannot be improved. I'd like to make it
more idiomatic... the lexer should be based on ranges, to begin
with. But it seems usable, and using it is what I want now.)

--HG--
branch : lmb/lua_like_config