Replaced interpolators' implementation with something different.
My previous code was based on Robert Penner's Action Script code,
which was licensed under a three-clause BSD license, which is not bad
by itself, but is different from the ZLib license, which used by
Allegro (and SDL 2, for that matter).

I'll probably licence FewDee under ZLib, too, just to be consistent
with the library I am using to do all the hard work. I was a bit
uneasy about that BSD-derived code "polluting" FewDee's codebase, so I
replaced it with code derived from Warren Moore's AHEasing
(https://github.com/warrenm/AHEasing), which is licensed under the "Do
What The Fuck You Want To Public License, Version 2".

The only drawback of Warren's code is that it does not provide the
"advanced" parameters for Elastic and Back interpolators. On the
bright side, my interpolators now have a completely homogeneous
interface; the interpolators example became simpler.

In some sense, the current implemention is closer to the spirit of
FewDee, which is strongly biased toward ease of use.