Fixed ConfigValue's initialization with literals '0' and '1'.
Those were causing the Boolean overload to be used. Now I have
templated constructors, one that accepts all numeric types, and one
that accepts only Booleans (but which is also a template, because
otherwise it would still be called for the literals -- non-template
overloads always have higher precedence).

Curiously, this was causing a unit test to fail. Was this something
that changed in recent versions of DMD or did I overlook a failing
test?!